+++ tools/talvos-cmd/CommandFile.h
@@ -33,14 +33,14 @@ private:
   void parseModule();
 
   template <typename T> void dump();
-  template <typename T> void fill(size_t Address, size_t NumBytes);
-  template <typename T> void range(size_t Address, size_t NumBytes);
+  template <typename T> void fill(uint64_t Address, uint64_t NumBytes);
+  template <typename T> void range(uint64_t Address, uint64_t NumBytes);
 
   std::istream &Stream;
   talvos::Device *Device;
   std::unique_ptr<talvos::Module> Module;
   const talvos::Function *Function;
-  std::map<std::string, std::pair<size_t, size_t>> Buffers;
+  std::map<std::string, std::pair<uint64_t, uint64_t>> Buffers;
   talvos::DescriptorSet DescriptorSet;
 
   std::string CurrentParseAction;
